feat(dpa4c): add compact invariant descriptor and CUDA deployment
Introduce DPA4C as a graph-native descriptor built from degree-scaled
Cartesian moments, exact invariant readouts, and pair-conditioned radial
modes.

- support backend-neutral training, serialization, graph export,
  calibration, and mixed-precision execution
- add compressed CUDA and canonical inference for the supported channel
  and angular profiles
- expose neighborhood masses, remove the fixed-capacity path, and tile
  compressed evaluation to bound memory
- cover parity, gradients, compression, export, and end-to-end
  energy/force/virial behavior

+++ source/api_c/include/c_api.h
@@ -442,7 +442,7 @@ extern void DP_DeepPotComputeEdgesGPUFloat32(DP_DeepPot* dp,
  * @param[in] nloc Number of owned local nodes.
  * @param[in] nall_nodes Total local-plus-halo node count.
  * @param[in] edge_storage Number of edge storage slots.
- * @since API version 28
+ * @since API version 29
  */
 extern void DP_DeepPotComputeCanonicalGraphGPU(
     DP_DeepPot* dp,
@@ -450,11 +450,11 @@ extern void DP_DeepPotComputeCanonicalGraphGPU(
     double* d_force,
     double* d_atom_virial,
     const int64_t* d_atype,
-    const int64_t* d_source,
+    const uint32_t* d_source,
     const float* d_edge_vec,
     const int64_t* d_destination_row_ptr,
     const int64_t* d_source_row_ptr,
-    const int64_t* d_source_order,
+    const uint32_t* d_source_order,
     int nloc,
     int nall_nodes,
     int64_t edge_storage);
