+++ source/api_cc/include/DeepPot.h
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 #pragma once
 
+#include <cstdint>
 #include <memory>
 
 #include "DeepBaseModel.h"
@@ -357,11 +358,11 @@ class DeepPotBackend : public DeepBaseModelBackend {
       double* d_force,
       double* d_atom_virial,
       const std::int64_t* d_atype,
-      const std::int64_t* d_source,
+      const std::uint32_t* d_source,
       const float* d_edge_vec,
       const std::int64_t* d_destination_row_ptr,
       const std::int64_t* d_source_row_ptr,
-      const std::int64_t* d_source_order,
+      const std::uint32_t* d_source_order,
       const int nloc,
       const int nall_nodes,
       const std::int64_t edge_storage);
@@ -829,11 +830,11 @@ class DeepPot : public DeepBaseModel {
                                    double* d_force,
                                    double* d_atom_virial,
                                    const std::int64_t* d_atype,
-                                   const std::int64_t* d_source,
+                                   const std::uint32_t* d_source,
                                    const float* d_edge_vec,
                                    const std::int64_t* d_destination_row_ptr,
                                    const std::int64_t* d_source_row_ptr,
-                                   const std::int64_t* d_source_order,
+                                   const std::uint32_t* d_source_order,
                                    const int nloc,
                                    const int nall_nodes,
                                    const std::int64_t edge_storage);
